# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  ok: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  username: string;
}

export interface NetStats {
  speedRx: string;
  speedTx: string;
  totalRx: string;
  totalTx: string;
}

export interface BotStats {
  status: string;
  number?: string | null;
  connectedAt?: string | null;
  uptime: number;
  memUsed: number;
  memTotal: number;
  memPercent: number;
  cpuLoad: string;
  platform: string;
  nodeVersion: string;
  userCount: number;
  fileCount: number;
  fileSizeMB: string;
  commandsToday: number;
  messagesHandled: number;
  net: NetStats;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  Connected: "Connected",
  Disconnected: "Disconnected",
  Connecting: "Connecting",
  Awaiting_QR_Scan: "Awaiting QR Scan",
  Logged_Out: "Logged Out",
  Error: "Error",
  Restoring: "Restoring",
  Initializing: "Initializing",
} as const;

export interface Session {
  id: string;
  label: string;
  number?: string | null;
  status: SessionStatus;
  connectedAt?: string | null;
  platform?: string | null;
  isMain: boolean;
  qr?: string | null;
  pairCode?: string | null;
}

export interface CreateSessionRequest {
  id: string;
  pairMode?: boolean;
  phone?: string;
}

export interface BroadcastRequest {
  message: string;
  targets?: string[];
}

export type BroadcastResultErrorsItem = {
  jid: string;
  error: string;
};

export interface BroadcastResult {
  sent: number;
  failed: number;
  total: number;
  errors: BroadcastResultErrorsItem[];
}

export interface BotSettings {
  botName?: string;
  prefix?: string;
  ownerNumber?: string;
  autoRead?: boolean;
  autoTyping?: boolean;
  nsfwEnabled?: boolean;
  premiumCode?: string;
  maintenanceMode?: boolean;
  maxSessions?: number;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  debug: "debug",
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  meta?: string | null;
}