# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WhatsApp Bot Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sessions:
    get:
      operationId: getSessions
      tags: [sessions]
      summary: Get all sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /logs:
    get:
      operationId: getLogs
      tags: [logs]
      summary: Get recent bot logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
      required:
        - token
        - username
...
    Session:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        number:
          type: string
          nullable: true
        status:
          type: string
          enum: [Connected, Disconnected, Connecting, "Awaiting QR Scan", "Logged Out", Error, Restoring, Initializing]
        connectedAt:
          type: string
          nullable: true
        platform:
          type: string
          nullable: true
        isMain:
          type: boolean
        qr:
          type: string
          nullable: true
        pairCode:
          type: string
          nullable: true
      required:
        - id
        - label
        - status
        - isMain
    CreateSessionRequest:
      type: object
      properties:
        id:
          type: string
        pairMode:
          type: boolean
        phone:
          type: string
      required:
        - id
...
    LogEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string
        meta:
          type: string
          nullable: true
      required:
        - id
        - timestamp
        - level
        - message