# lib/api-zod/src/generated/types/createSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSessionRequest {
  id: string;
  pairMode?: boolean;
  phone?: string;
}