# lib/api-zod/src/generated/types/logEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntryLevel } from "./logEntryLevel";

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  meta?: string | null;
}